move + simplify tag names
